// frontend_restaurants/web/src/lib/types.ts
export type StatisticChartType = {
	type: string;
	year: number;
	month?: number;
	week?: number;
	day?: number;
	timeOfDay: string;
};